Added processing feature to streams. This will be useful for thread synchronization.
git-svn-id: http://svn.modules.centreon.com/centreon-broker/trunk@929 7c7d615a-03b3-410d-a316-1135b1bd9499

+++ compression/test/stream/io.cc
@@ -112,7 +112,7 @@ int main() {
           bufferc = 0;
         }
         datac += cb;
-        if (datac == data->size())
+        if (datac == static_cast<unsigned int>(data->size()))
           data.clear();
       }
     }
